payment-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pymongo==4.6.0
pydantic==1.10.17
python-dotenv==1.0.1
PyJWT==2.8.0
stripe==11.4.1
